refactor: move logger, highlighter, and spinner to a layer-neutral ui home
The report layer imported three CLI modules for its logging, which made
report depend on the CLI front-end. They now live in src/ui/ with no
behavior change.
