+++ packages/nestjs-doctor/src/ui/highlighter.ts
