+++ packages/nestjs-doctor/src/ui/spinner.ts
